requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.9
httpx==0.27.2
chromadb==0.5.17
sentence-transformers==3.2.1
numpy==2.1.2
scipy==1.13.1
pypdf==5.1.0
trafilatura==1.12.2
google-generativeai==0.7.2
lxml_html_clean==0.3.1
